requirements.txt
matplotlib
numpy
pandas
tqdm
scikit-learn
torch 
torchvision 
torchstudio 
transformers
sentence-transformers
chromadb
whoosh
spacy
bert-score
rouge-score
openai
python-dotenv
nltk
